Move common_jdbc_methods to abstract/core to match other abstracted modules
